requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
PyPDF2>=3.0.0
requests>=2.31.0
transformers>=4.35.0
Pillow>=10.1.0
python-markdown>=3.5.1
beautifulsoup4>=4.12.2
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
